+++ CHANGELOG.md
@@ -10,6 +10,7 @@ glance - what to expact from upgrading to a new version.
 
 - Per-user and global rate limits.
 - `!hive` as alias to `!hive2`
+- Description of bot commands to help output.
 
 ### Changed